# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/shifts": {
      "post": {
        "operationId": "ShiftsController_createShift",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateShiftDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create Shift Schedule (Koordinator Only)",
        "tags": [
          "Shifts"
        ]
      },
      "get": {
        "operationId": "ShiftsController_getShifts",
        "parameters": [
          {
            "name": "month",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "year",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get All Shift List (Koordinator Only)",
        "tags": [
          "Shifts"
        ]
      }
    },
    "/shifts/bulk": {
      "post": {
        "operationId": "ShiftsController_createBulkShifts",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkCreateShiftDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Bulk Create Shift Schedule (Koordinator Only)",
        "tags": [
          "Shifts"
        ]
      }
    },
    "/shifts/my": {
      "get": {
        "operationId": "ShiftsController_getMyShifts",
        "parameters": [
          {
            "name": "month",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "year",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get My Personal Shift Schedule (Linmas Only)",
        "tags": [
          "Shifts"
        ]
      }
    },
    "/shifts/{shift_id}": {
      "put": {
        "operationId": "ShiftsController_updateShift",
        "parameters": [
          {
            "name": "shift_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateShiftDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update Shift Schedule (Koordinator Only)",
        "tags": [
          "Shifts"
        ]
      },
      "delete": {
        "operationId": "ShiftsController_deleteShift",
        "parameters": [
          {
            "name": "shift_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete Shift Schedule (Koordinator Only)",
        "tags": [
          "Shifts"
        ]
      }
    },
    "/shifts/{shift_id}/substitute": {
      "post": {
        "operationId": "ShiftsController_assignSubstitute",
        "parameters": [
          {
            "name": "shift_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubstituteShiftDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Assign Substitute Officer (Koordinator Only)",
        "tags": [
          "Shifts"
        ]
      }
    },
...
    "/storage/upload": {
      "post": {
        "operationId": "StorageController_uploadImage",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  },
                  "folder": {
                    "type": "string",
                    "enum": [
                      "attendance",
                      "incidents",
                      "patrols"
                    ]
                  },
                  "reference_id": {
                    "type": "string",
                    "nullable": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Upload Image to Backblaze B2 CDN",
        "tags": [
          "Storage"
        ]
      }
    },
    "/storage/files/{file_id}": {
      "get": {
        "operationId": "StorageController_getFileMetadata",
        "parameters": [
          {
            "name": "file_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get File Metadata",
        "tags": [
          "Storage"
        ]
      }
    },
    "/storage/files/{file_id}/view": {
      "get": {
        "operationId": "StorageController_getPrivateFileView",
        "parameters": [
          {
            "name": "file_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get Temporary Secure URL for Private Image",
        "tags": [
          "Storage"
        ]
      }
    },
    "/storage/files/batch-view": {
      "post": {
        "operationId": "StorageController_getMultiplePrivateFileViews",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GetMultipleFilesDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get Temporary Secure URLs for Multiple Private Images",
        "tags": [
          "Storage"
        ]
      }
    },
...
    "/salaries/recap": {
      "get": {
        "operationId": "SalariesController_getAllAdjustments",
        "parameters": [
          {
            "name": "month",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "year",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get Salary Adjustments Recap for All Linmas (Koordinator Only)",
        "tags": [
          "Salaries & Adjustments"
        ]
      }
    },
    "/salaries/my-recap": {
      "get": {
        "operationId": "SalariesController_getMyAdjustments",
        "parameters": [
          {
            "name": "month",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "year",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get My Salary Adjustments (Linmas Only)",
        "tags": [
          "Salaries & Adjustments"
        ]
      }
    },
    "/salaries/generate": {
      "post": {
        "operationId": "SalariesController_generate",
        "parameters": [
          {
            "name": "month",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "year",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Generate and persist salary adjustments",
        "tags": [
          "Salaries & Adjustments"
        ]
      }
    },
    "/salaries/saved-recap": {
      "get": {
        "operationId": "SalariesController_getSavedRecap",
        "parameters": [
          {
            "name": "month",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "year",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get persisted salary adjustments",
        "tags": [
          "Salaries & Adjustments"
        ]
      }
    },
    "/salaries/{periodId}/approval": {
      "patch": {
        "operationId": "SalariesController_approveSalary",
        "parameters": [
          {
            "name": "periodId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApproveSalaryAdjustmentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Approve or revoke salary adjustment",
        "tags": [
          "Salaries & Adjustments"
        ]
      }
    },
...
  },
  "info": {
    "title": "MANALINMAS API",
    "description": "Manajemen Kinerja Linmas & Pelaporan Darurat API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateShiftDto": {
        "type": "object",
        "properties": {
          "shift_date": {
            "type": "string",
            "example": "2026-05-21"
          },
          "shift_type": {
            "type": "string",
            "enum": [
              "MORNING",
              "NIGHT"
            ],
            "example": "NIGHT"
          },
          "regu_id": {
            "type": "string",
            "example": "uuid-of-regu",
            "description": "If provided, auto-assigns all members of this Regu"
          },
          "assigned_linmas_ids": {
            "example": [
              "LMS-001",
              "LMS-002"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "start_time": {
            "type": "string",
            "example": "19:00:00",
            "description": "HH:MM:SS format"
          },
          "end_time": {
            "type": "string",
            "example": "07:00:00",
            "description": "HH:MM:SS format"
          }
        },
        "required": [
          "shift_date",
          "shift_type",
          "start_time",
          "end_time"
        ]
      },
      "BulkCreateShiftDto": {
        "type": "object",
        "properties": {
          "shifts": {
            "description": "Array of shifts for the month",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateShiftDto"
            }
          }
        },
        "required": [
          "shifts"
        ]
      },
      "UpdateShiftDto": {
        "type": "object",
        "properties": {
          "shift_type": {
            "type": "string",
            "enum": [
              "MORNING",
              "NIGHT"
            ],
            "example": "MORNING"
          },
          "regu_id": {
            "type": "string",
            "example": "uuid-of-regu"
          },
          "assigned_linmas_ids": {
            "example": [
              "LMS-003"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "SubstituteShiftDto": {
        "type": "object",
        "properties": {
          "original_linmas_id": {
            "type": "string",
            "example": "uuid-of-original-linmas"
          },
          "substitute_linmas_id": {
            "type": "string",
            "example": "uuid-of-substitute-linmas"
          },
          "reason": {
            "type": "string",
            "example": "Sakit / Izin"
          }
        },
        "required": [
          "original_linmas_id",
          "substitute_linmas_id",
          "reason"
        ]
      },
...
      "GetMultipleFilesDto": {
        "type": "object",
        "properties": {
          "file_ids": {
            "example": [
              "uuid-file-1",
              "uuid-file-2"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "file_ids"
        ]
      },
...
      "ApproveSalaryAdjustmentDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}